Conflicts:
	DollarBills.xcodeproj/project.pbxproj
	DollarBills/Extensions/ViewModelExt.swift
	DollarBills/Logic and Framework/Supabase.swift
	DollarBills/Views/MainView.swift

+++ DollarBills/Component/CustomContent.swift
@@ -229,7 +229,7 @@ struct VerticalProgressBar: View {
 
 struct LeaderboardBar: View {
     
-    var points: Double
+    var points: Int
     var rank: Int
     var name: String
     var body: some View {
